Address remote shell parser review
